+++ hdc.py
